backend/pyproject.toml
[project]
name = "cash-flow-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "uvicorn>=0.37.0",
]
